assertion fails
