+++ src/it/correct/src/main/java/io/github/eocqrs/simple/Simple.java
